moved a component
